hack: stub out sbrk
This is a demo of how you can stub out sbrk to test things.

You only need to do one more thing: in envsetup, fake up some
directory entries as from a real dirreadall and then you can test your code.
The way you do that is to create a static array of the structs and instead of doing a
dirreadall just set the point to the structs (just make an array of 3 or so) and set nd to 3.

Another way? Create a program that runs on harvey that does a dirreadall of #e and then prints
the C code for the static initialized struct to a file. Then just include that static initializer
in pdksh.

Not super hard, and it lets you debug on linux with gdb, which will save you a lot of time.

This is actually commit ready.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>