Debug alloc
I replaced alloc with malloc (see noalloc.c) and we get to a pdksh prompt.

IF YOU SET HACK
With this change, you can debug pdksh on linux.

The failure seems to be the same as on Harvey.

gdb pdksh
break afree
run

have fun. It fails in the same manner it fails on Harvey.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>