Remove more personal copyrights
Now that we have an official HarveyOS organization, we can have a more formal copyright policy.
The policy from here on out is as follows:

we don't remove personal copyrights from imported code
when we create code, the copyright is HarveyOS
we don't add personal copyrights to code
if needed, when substantial changes are made, we'll add a HarveyOS copyright notice.
We expect this to be infrequent.
If we find other places where we have added personal copyrights since the
Harvey project began, we will remove them as well.

This policy accords with modern practice on projects such as Harvey.

Signed-off-by: Álvaro Jurado <elbingmiss@gmail.com>