add a panic function. Panic on sigsuspend, which is broken
This at least diagnoses PART of the problems with pdksh, in that
it was spinning on unchanging memory variables and never bothering
to actually use a system call to see if the command had exited.
That's why it would "hang" on the command exiting. It was not
hanging. It was not ever checking.

It's a good idea, with unimplemented functions, to call panic.

If we had gdbserver we would probably have found this 6 months ago
at least.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>